wip creating and updating

+++ back/src/main/java/com/example/demo/ChoreRepository.java
@@ -2,6 +2,6 @@ package com.example.demo;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChoreRepository extends CrudRepository<ChoreEntity, Integer> {
+public interface ChoreRepository extends CrudRepository<ChoreEntity, Long> {
     
 }
